requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Video Processing
av>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
colorama>=0.4.6

# Monitoring and Profiling
psutil>=5.9.0
py-cpuinfo>=9.0.0

# API (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Data handling
pandas>=2.0.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0
pylint>=2.17.0

# Development
ipython>=8.14.0
jupyter>=1.0.0
